bot/requirements.txt
aiogram
watchdog
python-dotenv
tinydb